Move GameBoard and GuessRow to components lib folder
